update new file riska

+++ admin/produk/index.php
@@ -12,8 +12,7 @@ include_once("../config/config.php");
                         <h3 class="card-title">Data Produk</h3>
                         <div class="card-tools">
                             <!-- This will cause the card to maximize when clicked -->
-                            <a href='produk/create.php?page=produk' class="btn btn-info"><i
-                                    class="fas fa-plus"></i>Tambah Produk</a>
+                            <a href='produk/create.php?page=produk' class="btn btn-info"><i class="fas fa-plus"></i>Tambah Produk</a>
                         </div>
                         <!-- /.card-tools -->
                     </div>
@@ -21,7 +20,7 @@ include_once("../config/config.php");
                         <table width='100%' id='example2' class="table table-bordered table-striped">
                             <thead>
                                 <tr>
-                                    <th>No</th>
+                                    <th style="width: 10px">No</th>
                                     <th>Nama Produk</th>
                                     <th>Kategori</th>
                                     <th>Deskripsi</th>
@@ -43,9 +42,9 @@ include_once("../config/config.php");
                             $result = mysqli_query($mysqli, $query);
                             if (mysqli_num_rows($result) > 0) {
                                 $no = 1; // Nomor urut
-                            
+
                                 while ($data = mysqli_fetch_array($result)) {
-                                    ?>
+                            ?>
                                     <tbody>
                                         <tr>
                                             <td>
@@ -69,21 +68,19 @@ include_once("../config/config.php");
                                             </td>
                                             <td><img src="produk/image/<?= $data['image'] ?>" width="100"></td>
                                             <td>
-                                                <a class="btn btn-success"
-                                                    href='produk/update.php?id=<?= $data['id'] ?>&page=produk'>Edit</a>
-                                                <a class="btn btn-danger" onclick='return confirmDelete()'
-                                                    href='produk/delete.php?id=<?= $data['id'] ?>&page=produk'>Hapus</a>
+                                                <a class="btn btn-success" href='produk/update.php?id=<?= $data['id'] ?>&page=produk'>Edit</a>
+                                                <a class="btn btn-danger" onclick='return confirmDelete()' href='produk/delete.php?id=<?= $data['id'] ?>&page=produk'>Hapus</a>
                                             </td>
                                         </tr>
                                     </tbody>
-                                    <?php
+                                <?php
                                 }
                             } else {
                                 ?>
                                 <tr>
                                     <td colspan="8">Tidak ada data produk.</td>
                                 </tr>
-                                <?php
+                            <?php
                             }
                             ?>
